moved all the recycle code into its own module
